First commit of Resource system
Added zlib libs and includes.
Added tinyxml lib.
Fixed a bug with AllocMap.
Added DirectoryResourceSource class, reads contents of a directory.
Added ZipResourceSource class, reads contents of a zip file.
Added MasterDirectoryResourceSource class.
	Uses DirectoryResourceSource and ZipResourceSource to access
	contents of subdirectories and zip files. Makes all of the
	contents available, good for use with modding. Mod authors can
	easily drop a new zip file in a directory and it's assets will be
	made available.
Added IResourceLoader.h, still incomplete but provides a translation layer
for caches to load in new resources.
Added ResourceCache class, untested but designed to maintain a copy of
resources so in memory with reasonable intelligence so that further
requests for those resources don't hit the hard drive.
Added ResourceHandle class, actually responsible for maintaining the
life-cycle of a resource for the ResourceCache